Estilizando + adicionando algumas funcionalidades

+++ src/pages/Home.jsx
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
 import Categories from '../components/Categories';
 import ProductInfo from '../components/ProductInfo';
 import { getProductsFromCategoryAndQuery } from '../services/api';
@@ -27,7 +28,6 @@ class Home extends Component {
   fetchProductByCategory = async (event) => {
     const { target } = event;
     const categorieSelected = target.value;
-
     const dataProduct = await getProductsFromCategoryAndQuery(categorieSelected, '');
     this.setState({ productList: dataProduct });
   };
@@ -38,6 +38,15 @@ class Home extends Component {
       const products = JSON.parse(cartlist);
       if (products.some((product) => product.id === newProduct.id)) {
         const index = products.findIndex((product) => product.id === newProduct.id);
+        if (products[index].available_quantity < products[index].productQuantity + 1) {
+          Swal.fire({
+            title: 'Error',
+            text: 'Não há mais unidades desse produto disponíveis',
+            icon: 'error',
+            confirmButtonText: 'Okay!',
+          });
+          return;
+        }
         products[index].productQuantity += 1;
         localStorage.setItem('cartlist', JSON.stringify(products));
       } else {
@@ -65,44 +74,51 @@ class Home extends Component {
     const { product, productList, totalItemsCart } = this.state;
     return (
       <div>
-        <input
-          type="text"
-          name="product"
-          value={ product }
-          data-testid="query-input"
-          onChange={ this.inputText }
-        />
-        <p
-          data-testid="home-initial-message"
-        >
-          Digite algum termo de pesquisa ou escolha uma categoria.
-        </p>
-
-        <button
-          type="button"
-          data-testid="query-button"
-          onClick={ this.fetchProduct }
-        >
-          Pesquisar
-        </button>
-
-        <button type="button">
-          <Link to="/shopping-cart" data-testid="shopping-cart-button">
-            Carrinho de Compras
-            <br />
-          </Link>
-          <span data-testid="shopping-cart-size">{ totalItemsCart }</span>
-        </button>
-        <Categories
-          fetchProductByCategory={ this.fetchProductByCategory }
-        />
-        <ProductInfo
-          productList={ productList }
-          addToCart={ this.addToCart }
-        />
+        <header>
+          <div>
+            <div className="header--search">
+              <input
+                type="text"
+                className="header--search--input"
+                name="product"
+                value={ product }
+                data-testid="query-input"
+                onChange={ this.inputText }
+              />
+              <button
+                type="button"
+                data-testid="query-button"
+                className="header--search--btn"
+                onClick={ this.fetchProduct }
+              >
+                Pesquisar
+              </button>
+            </div>
+            <p
+              data-testid="home-initial-message"
+              className="header--title"
+            >
+              Digite algum termo de pesquisa ou escolha uma categoria.
+            </p>
+          </div>
+          <h1>Online Store</h1>
+          <div className="cart-button" type="button">
+            <Link to="/shopping-cart" data-testid="shopping-cart-button">
+              <span data-testid="shopping-cart-size">{ totalItemsCart }</span>
+            </Link>
+          </div>
+        </header>
+        <div className="home-container">
+          <Categories
+            fetchProductByCategory={ this.fetchProductByCategory }
+          />
+          <ProductInfo
+            productList={ productList }
+            addToCart={ this.addToCart }
+          />
+        </div>
       </div>
     );
   }
 }
-
 export default Home;
